web/data/e739c7c5-a2e2-42e4-bfad-4c2dccdedefe/requirements.txt
PyYAML
gradescope-utils